+++ src/App.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import logo from './logo.svg';
+import './App.css';
+
+function App() {
+  return (
+          <><p>
+      Hello World.
+    </p><a
+      className="App-link"
+      href="https://reactjs.org"
+      target="_blank"
+      rel="noopener noreferrer"
+    >
+      </a></>
+  );
+}
+
+export default App;
